+++ libsolidity/codegen/ir/IRGeneratorForStatements.cpp
@@ -188,6 +188,23 @@ private:
 			solAssert(suffix == "offset" || suffix == "length", "");
 			value = IRVariable{*varDecl}.part(suffix).name();
 		}
+		else if (
+			auto const* functionType = dynamic_cast<FunctionType const*>(varDecl->type());
+			functionType && functionType->kind() == FunctionType::Kind::External
+		)
+		{
+			solAssert(suffix == "selector" || suffix == "address", "");
+			solAssert(varDecl->type()->sizeOnStack() == 2, "");
+			if (suffix == "selector")
+				// Left align the 4 byte selector
+				value =
+					m_context.utils().shiftLeftFunction(256 - 32) +
+					"(" +
+					IRVariable{*varDecl}.part("functionSelector").name() +
+					")";
+			else
+				value = IRVariable{*varDecl}.part("address").name();
+		}
 		else
 			solAssert(false, "");